resolve `rootFolder` issue
